+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+### alpha 0.4.20251201
+- minor bugfixes and refactoring
+- added Qwen3-VL (loads to CPU - slow)
+  - Gets pictures from browser extension or phone camera.
+  - Removed previous (javascript) barcode scanner; Now uses Qwen3 as well (~1.5s on my CPU).
+
 ### alpha 0.3.20251129
 - fixed a bug with shopping/todo lists
-- improved visual feedback and voice in recording window
+- improved visual feedback and voice-in recording window
 - made the chat window a standard rolling window 
   - chat session now remembers up to context size and forgets older stuff instead of throwing an error
 - automated local ssl certificate generation and refresh